Introduce a way for entries generated by tests to be associated with their related function declaration (#3509)
closes #3508.

To continue with improving in-language testing (like the linked issues),
we need to have a way of getting the original function declaration for
the test that we are running from `forc-test`. Currently `forc-test`
only works on `FinalizedEntry`s generated by the tests and we cannot get
the function declaration from the entry point. Here I am introducing an
optional field to the `FinalizedEntry` for tests. If the
`FinalizedEntry` is generated from a test function, corresponding
`FinalizedEntry`'s declaration id is populated and can be retrieved. To
pass this declaration id information I needed to preserve it in the IR
generation step to do that I added a metadata for it and populate it for
the test function declarations. So for test function declarations we are
inserting their declaration index to the metadata. I did not insert the
`DeclarationId` since it does not implement `Hash`, and we are already
inserting the span for the function declaration. Since `DeclarationId`
is basically the index + span, storing only the index was enough. After
the IR generation before creating an entry point I construct the
`DeclarationId` from index and span (which comes from the metadata) and
pass that to `FinalizedEntry`. This way in the IR generation step we are
not losing the declaration index + span (and thus, the declaration id)
and can retrieve the test function's declaration later on.

unblocks #3490.
unblocks #3266.