Add a way to specify salt while declaring contract dependencies (#3531)
closes #2970.

## Summary
This PR introduces a way of declaring `Salt` with contract dependencies.
To specify a salt for a contract dependency, users can basically provide
the salt next to the source of the dependency. Example:

```toml
[contract-dependencies]
foo = { path = "../foo", salt = "0x11111111111111111111111111111111111111111111111111111111111111111" }
```
Declared salts are taken into account while injecting the contract
dependency ids into the dependee's namespace.

## Implementation Details
I am adding the `Salt` information to the edges as conceptually they are
only meaningful between two nodes in the graph (it is a property of the
dependency relation just like dependency type). I started implementing
this with Salt being part of the `Pinned` package, but that creates
situations where two different salt declarations for the same package
creates two different nodes in the graph and create double compilations
so I decided to add them to the edges as it both solves double
compilations and made more sense conceptually.

I added an optional field for `Salt` in the edges of the graph which
gets populated only if the dependency kind described by that edge is
`DepKind::Contract` (a contract dependency). For contract dependencies
declared under `[contract-dependencies]` table in `Forc.toml`, we have
the ability of specifying the salt. If no salt is provided
`Salt::zeroed` is used instead. An example for the created dependency
graph can be seen below:

![success_case](https://user-images.githubusercontent.com/20915464/205971943-f47156e1-9813-4dee-b928-ae26f9751231.png)


I am also checking for conflicting salt declarations, meaning that if
two different contracts that depends on the same contract declares
different salt for that contract, I am producing an error. An example
dependency graph for that case can be seen below (see the edges going to
contract_c):

![error_case](https://user-images.githubusercontent.com/20915464/205971390-943b913c-fb5d-4f0c-ac2d-1360ffed083a.png)

Tests for both cases are added. Graphs seems a little small but can be
expanded if clicked on it. (I wonder if there is a good solution for
hosting dot output with markdown here 😄)